Created packets to sync tabbed inventories
These tabbed screens are initiated from the client side, as opposed to the server side. This makes it difficult to initiate screenHandler syncing, because it's usually taken care of on the server side. I needed to make packets to alert the server with the syncID and for the serverPlayerEntity to attach listeners to the proper ScreenHandler.